functions/basic/pubsub-trigger/src/requirements.txt
functions-framework==3.7.0
google-cloud-storage==3.4.1
